+++ public/js/updateMenuItem.js
@@ -23,21 +23,28 @@ const updateMenuItem = async(event) => {
 
 // Functionality for Delete button as well
 const deleteMenuItem = async(event) => {
-  if (event.target.hasAttribute('data-id')) {
-    const id = event.target.getAttribute('data-id');
-    const response = await fetch(`/api/menu/deleteMenuItem/${id}`, {
-      method: 'DELETE',
-    });
-    
-    if (response.ok) {
-      alert('Are you sure you want to delete this menu item and all things related to it? This action cannot be undone.');
-      document.location.replace('/menu');
-    } 
-    else {
-      alert(response.statusText);
+  if (confirm("Are you sure you want to delete this menu item and all things related to it? This action cannot be undone!!") == true) {
+    if (event.target.hasAttribute('data-id')) {
+      const id = event.target.getAttribute('data-id');
+      const response = await fetch(`/api/menu/deleteMenuItem/${id}`, {
+        method: 'DELETE',
+      });
+      
+      if (response.ok) {
+        // alert('Are you sure you want to delete this menu item and all things related to it? This action cannot be undone.');
+        document.location.replace('/menu');
+      } 
+      else {
+        alert(response.statusText);
+      };
     };
-  };
+  }
+  else 
+  {
+    document.location.replace('/menu');
+  }
 };
 
 document.querySelector('#updateMenuItem').addEventListener('submit', updateMenuItem);
+
 document.querySelector('#deleteButton').addEventListener('click', deleteMenuItem)
